@@ -1,13 +1,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 32
 
     defaultConfig {
         applicationId "com.weighing.scale"
-        minSdkVersion 21
-        versionCode 1
-        versionName "1.0"
+        minSdkVersion 23
+        targetSdk 32
+        versionCode 2
+        versionName "2.0"
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
 
@@ -18,15 +19,18 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        targetCompatibility JavaVersion.VERSION_11
+        sourceCompatibility JavaVersion.VERSION_11
+    }
+    buildToolsVersion buildToolsVersion
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-
-    implementation 'com.google.android.gms:play-services-location:18.0.0'
-    // butter knife
-    /*implementation 'com.jakewharton:butterknife:8.8.1'*/
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-   /* annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'*/
+    implementation 'com.google.android.material:material:1.8.0'
+    implementation 'com.google.android.gms:play-services-location:21.0.1'
     implementation files('libs/jxl.jar')
+    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0'
+
 }
